requirements.txt
# Restaurant Dashboard - iFood API Version
# Python dependencies

# Web Framework
flask>=2.3.0

# Database
psycopg2-binary>=2.9.0

# Password hashing
bcrypt>=4.0.0

# HTTP requests for iFood API
requests>=2.31.0

# Date/time handling (optional but useful)
python-dateutil>=2.8.0

# Environment variables (optional)
python-dotenv>=1.0.0
